[COGB-53] Criar classe construtora GetQuerySQL
## [COGB-53] Descrição
Criei a classe construtora de query get que estende a QuerySQL padrão
